refactored task user methods to traits
